Move header to separate component
