+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "initiative-tracker",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
