+++ src/components/Header.css
@@ -1,0 +1,44 @@
+.options-container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0.5rem;
+}
+
+.form-check-input {
+  width: 2.25rem;
+  height: 1.25rem;
+  background-color: var(--btn-bg);
+  border-radius: 1rem;
+  border: 1px solid var(--btn-border);
+  transition: background-color 0.3s ease;
+  cursor: pointer;
+  appearance: none;
+  position: relative;
+}
+
+.form-check-input:checked {
+  background-color: var(--btn-bg-active);
+}
+
+.form-check-input::before {
+  content: "";
+  position: absolute;
+  top: 2px;
+  left: 2px;
+  width: 1rem;
+  height: 1rem;
+  background-color: var(--btn-toggle);
+  border-radius: 50%;
+  transition: transform 0.3s ease;
+}
+
+.form-check-input:checked::before {
+  transform: translateX(1rem);
+}
+
+.form-check-label {
+  margin-left: 0.5rem;
+  color: var(--btn-text);
+  user-select: none;
+}
